requirements.txt
piper-tts>=1.2.0
sounddevice>=0.4.6
numpy>=1.24.0
keyboard>=0.13.5
pathlib
